Updates for newer kernels
While testing with 5.16 there are some changed
struct members that need updates. Most notably
the change to task_struct.__state breaks in a few places
